VMS-61121: Replace "WERROR_IF LINUX" with WERROR
Related to VMS-61121

+++ artifacts/udt/src/udt/connection_buffer.cpp
@@ -175,7 +175,7 @@ int SendBuffer::addBufferFromFile(fstream& ifs, int len)
             pktlen = m_iMSS;
 
         ifs.read(s->m_pcData, pktlen);
-        if ((pktlen = ifs.gcount()) <= 0)
+        if ((pktlen = (int) ifs.gcount()) <= 0)
             break;
 
         // currently file transfer is only available in streaming mode, message is always in order, ttl = infinite
