+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/common/widgets/dropdown_button.cpp
@@ -159,7 +159,7 @@ QSize DropdownButton::sizeHint() const
         return base_type::sizeHint();
 
     Q_D(const QPushButton);
-    if (d->sizeHint.isValid() && d->lastAutoDefault == autoDefault())
+    if (d->sizeHint.isValid() && d->lastAutoDefault == (unsigned int) autoDefault())
         return d->sizeHint;
 
     d->lastAutoDefault = autoDefault();
