+++ vms/client/nx_vms_client_desktop/src/ui/graphics/instruments/selection_fixup_instrument.cpp
@@ -35,8 +35,8 @@ protected:
 
 private:
     void fixButtonDownPositions(QGraphicsSceneMouseEvent *event) {
-        for (uint i = 0x1; i <= 0x10; i <<= 1) {
-            if((event->buttons() & i) || event->button() == i)
+        for (int i = 0x1; i <= 0x10; i <<= 1) {
+            if((event->buttons() & i) || (int) event->button() == i)
                 continue;
 
             /* Qt is totally inconsistent in setting the button down positions for
